refactor: moved context.ts into utlis
