+++ IndoorMapsBackend/src/utils/seed.ts
@@ -1,6 +1,6 @@
 // currently this file is only used in the buildings.test.ts, it ensures that there is a user that can own the building
 
-import { prisma } from '../context.js'
+import { prisma } from './context.js'
 export async function seed() {
     const seedUser = await prisma.user.upsert({
         where: { email: 'seed@test.com' },
